% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{add_maandnaam}
\alias{add_maandnaam}
\title{Toevoegen van de maandnaam}
\usage{
add_maandnaam(df, datum = "datum", titlecase = FALSE)
}
\arguments{
\item{df}{Dataframe waar een kolom maandnaam aan wordt toegevoegd}

\item{datum}{Character met de naam van de datum kolom. De default is \code{"datum"}}

\item{titlecase}{Logical. Maandnamen met hoofdletter of niet}
}
\value{
Een dataframe met de nieuwe kolom \code{maandnaam}
}
\description{
Deze functie voegt de Nederlandse naam van een maand toe aan een dataframe op basis van de datum
}
\examples{
\dontrun{

dataframe <- add_maandnaam(dataframe_orig)
dataframe \%>\% add_maandnaam()
dataframe \%>\% add_maandnaam(datum = "mijn_datumkolom", titlecase = FALSE)
}
}
