% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_biologie_utils.R
\name{add_twn_voorkeurnaam}
\alias{add_twn_voorkeurnaam}
\title{Toevoegen TWN-voorkeurnamen}
\usage{
add_twn_voorkeurnaam(df, naam_orig = naam, naam_doel = naam_voorkeur)
}
\arguments{
\item{df}{Een dataframe met biologische data}

\item{naam_orig}{Naam van de kolom met originele namen. Dit argument is
quoted. Default is \code{naam}}

\item{naam_doel}{Naam van de kolom waar de voorkeurnamen in moeten komen. Dit
argument is quoted. Default is \code{naam_voorkeur}}
}
\value{
Een dataframe met daaraan de voorkeurnamen toegevoegd.
}
\description{
Deze functie voegt op basis van de TWN-lijst de voorkeurnamen toe aan een
dataframe. Deze functie maakt gebruik van de dataset twn_lijst uit het package autecodata.
}
\note{
Deze functie is afhankelijk van de beschikbaarheid van de package autecodata. Dit kan gedownload worden op \url{https://github.com/RedTent/autecodata}

Bij een ongeldige TWN-naam geeft deze functie een \code{NA}.

De namen zijn hoofdlettergevoelig.
}
\examples{
\dontrun{

bio \%>\% add_twn_voorkeurnaam(naam_orig = naam, naam_doel = naam_voorkeur)

}

}
\seealso{
\link[autecodata:twn_lijst]{autecodata::twn_lijst}
}
