% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_kaarten.R
\name{basiskaart}
\alias{basiskaart}
\title{Leaflet basiskaart met topo en luchtfoto}
\usage{
basiskaart(data = NULL, type = c("osm", "cartolight"), api_key = NULL, ...)
}
\arguments{
\item{data}{Een optioneel dataframe. Zie ook het data-argument in de functie \code{\link[leaflet:leaflet]{leaflet::leaflet()}}}

\item{type}{Het type kaart. Opties zijn \code{"osm"} (default) en \code{"cartolight"}}

\item{api_key}{API key voor Carto achtergrond. Zonder key krijgt de achtergrond een watermerk.
Indien NULL wordt \code{Sys.getenv("CARTO_API_KEY")} geraadpleegd.}

\item{...}{Named arguments voor \code{leaflet()}}
}
\value{
Een leaflet kaart
}
\description{
Deze functie genereert een basiskaart in leaflet. Aan deze kaart kunnen vervolgens vrij elementen worden toegevoegd
}
\examples{

basiskaart()

}
