% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{boxplot_basis}
\alias{boxplot_basis}
\title{Boxplots per jaar}
\usage{
boxplot_basis(data, mp = NULL, mpomsch = NULL, parnaam = NULL,
  eenheid = NULL, ...)
}
\arguments{
\item{data}{Een dataframe met de gegevens van 1 tijdreeks, dus van 1 meetpunt en 1 parameter.
Kolommen zoals beschreven in \code{\link[=import_fys_chem]{import_fys_chem()}}.}

\item{mp}{Code van het meetpunt. Deze wordt gebruikt in de titel}

\item{mpomsch}{Omschrijving van het meetpunt. Deze wordt gebruikt in de titel}

\item{parnaam}{Naam van de parameter. Deze wordt gebruikt in de ondertitel}

\item{eenheid}{Eenheid van de parameter. Deze wordt gebruikt als titel van de Y-as}

\item{...}{Heeft geen functie, maar kan gebruikt worden om overbodige parameters te negeren}
}
\value{
Een ggplot boxplot grafiek. Het is mogelijk om achteraf andere ggplot objecten toe te voegen met \code{+}
}
\description{
Deze functie plot een boxplot per jaar van 1 meetpunt van 1 parameter.
De functie dient eigenlijk alleen met complete meetjaren te worden gebruikt.
}
\examples{
\dontrun{

## ERROR

}
}
