% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_get_data.R
\name{brand_hhsk}
\alias{brand_hhsk}
\alias{get_brand_hhsk}
\alias{get_template_hhsk}
\title{De quarto brand-extensie voor HHSK}
\usage{
get_brand_hhsk()

get_template_hhsk()
}
\value{
Installeert de brand extensie in de working directory
}
\description{
Installeert de brand-extensie van HHSK of geeft instructies voor het installeren
van de templates
}
\examples{

\dontrun{
 get_brand_hhsk()
 get_template_hhsk()
}

}
