% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{coordinate_conversion}
\alias{coordinate_conversion}
\title{Coordinate Conversion}
\usage{
coordinate_conversion(x, y, source_crs = "+init=EPSG:28992",
  goal_crs = "+init=EPSG:4326")
}
\arguments{
\item{x}{x-coordinaat van de bron}

\item{y}{y-coordinaat van de bron}

\item{source_crs}{Coordinaten Referentie Systeem van de bron coordinaten. Default is RD-stelsel.}

\item{goal_crs}{Coordinaten Referentie Systeem van de doelcoordinaten. Default is WGS84.}
}
\value{
Een dataframe met een long en een lat
}
\description{
Helper functie voor coordinaten conversie. Deze functie wordt gebruikt door add_lat_long
}
\examples{
\dontrun{

coordinate_conversion(x = 111111, y = 444444)
}
}
