% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{count_rel}
\alias{count_rel}
\title{Relative count - fractions}
\usage{
count_rel(x, ..., sort = FALSE, name = "n")
}
\arguments{
\item{x}{a \code{tbl()} to count.}

\item{...}{Variables to group by.}

\item{sort}{if TRUE will sort output in descending order of n}

\item{name}{The output column name. If omitted, it will be n.}
}
\value{
A tbl with fractions, grouped the same way as x.
}
\description{
Relative count - fractions
}
\examples{

count_rel(mtcars, cyl, sort = TRUE)
count_rel(mtcars, cyl)

}
