% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{data_online}
\alias{data_online}
\title{Download en lees een online RDS-bestand}
\usage{
data_online(file, url = NULL)
}
\arguments{
\item{file}{Bestandsnaam van het rds bestand om te downloaden en in te lezen.}

\item{url}{URL van de bronlocatie. Default is de FTP-server van HHSK}
}
\value{
Een R-object
}
\description{
Deze functie download een .rds bestand van de FTP-server en leest het in.
}
\examples{

\dontrun{
fys_chem <- data_online("fys_chem.rds")
}
}
