% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dawaco_connect.R
\name{dawaco_connect}
\alias{dawaco_connect}
\title{Open een verbinding naar de Dawaco-database}
\usage{
dawaco_connect(gebruiker = "behtent", wachtwoord_env = "DAWACO2018_PWD")
}
\arguments{
\item{gebruiker}{Gebruikersnaam}

\item{wachtwoord_env}{Environment variabele met het wachtwoord. Als deze niet bestaat wordt er een
venster geopend om het wachtwoord in te voeren.}
}
\value{
Een connectie naar de Dawaco-database
}
\description{
Open een verbinding naar de Dawaco-database
}
\examples{

\dontrun{

con <- dawaco_connect()

meetpunten <- dplyr::tbl(con, dbplyr::in_schema("DAWACO2018", "MP"))

}
}
