% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{download_data}
\alias{download_data}
\title{Download data van FTP-server
Download data van FTP-server
Download data van FTP-server}
\usage{
download_data(files, destination = "data", url = NULL)
}
\arguments{
\item{files}{Een character-vector met bestandsnamen}

\item{destination}{Map waar de data opgeslagen moet worden.}

\item{url}{URL van de FTP-server}
}
\value{
Databestanden op de gekozen locatie.
}
\description{
Download data van FTP-server
Download data van FTP-server
Download data van FTP-server
}
\examples{
\dontrun{
download_data(c("fys_chem.rds", "meetpunten.rds"))
}

}
