% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_rapportage_utils.R
\name{dt_labels_nederlands}
\alias{dt_labels_nederlands}
\title{Nederlandse labels voor Datatable}
\usage{
dt_labels_nederlands()
}
\value{
Aanpassing in \code{\link[base:options]{base::options()}}
}
\description{
Deze functie stelt in \code{\link[base:options]{base::options()}} een aantal opties in die zorgen voor Nederlandse labels
bij het gebruik van \link[DT:datatable]{DT::datatable}
}
\examples{

dt_labels_nederlands

}
