% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{fys_chem}
\alias{fys_chem}
\title{Laad fys_chem.rds}
\usage{
fys_chem(fys_chem_rds = NULL)
}
\arguments{
\item{fys_chem_rds}{Pad naar de locatie van het .rds-bestand met fysisch-chemische metingen.
Indien \code{NULL} wordt achtereenvolgens gezocht naar fys_chem.rds in de map data en in de map
P:/Dawaco/FME}
}
\value{
Het ingelezen object uit fys_chem.rds
}
\description{
Deze functie leest fys_chem.rds om fysisch chemische metingen sneller in te kunnen lezen.
}
\examples{

\dontrun{

data <- fys_chem()
}
}
