% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_get_data.R
\name{get_logo}
\alias{get_logo}
\title{Logo's HHSK}
\usage{
get_logo(pad = "images")
}
\arguments{
\item{pad}{Het pad op de logo's op te slaan. Wordt aangemaakt als het nog niet bestaat. Default is de map \emph{images}.}
}
\value{
slaat logo_kleur.png en logo_wit.png op op de gekozen locatie.
}
\description{
Logo's van HHSK als bestand
}
\examples{

\dontrun{
get_logo("images")
}

}
