% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_get_data.R
\name{get_open_gisdata}
\alias{get_open_gisdata}
\title{Haal Open GIS-data van HHSK op}
\usage{
get_open_gisdata(dataset = NULL)
}
\arguments{
\item{dataset}{Code voorde gewenste dataset. Zie ook \code{gisdata_info}. \code{NULL} geeft een keuze menu.}
}
\value{
Een geo-object met de gekozen dataset
}
\description{
Met deze functie kan GIS-data van HHSK worden gedownload. Voor de beschikbare datasets zie
\code{gisdata_info} of gebruik de interactieve versie van de functie met \code{dataset = NULL}.
}
\examples{

\dontrun{
get_open_gisdata("krw_2022_2027")
}

}
