% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_get_twn_info.R
\name{get_twn_info}
\alias{get_twn_info}
\alias{get_twn_voorkeur}
\alias{get_twn_parent}
\alias{get_twn_status}
\alias{get_twn_localname}
\alias{get_twn_taxonlevel}
\title{Opzoeken van de TWN informatie}
\usage{
get_twn_voorkeur(namen)

get_twn_parent(namen)

get_twn_status(namen)

get_twn_localname(namen)

get_twn_taxonlevel(namen)
}
\arguments{
\item{namen}{Een vector met taxonnamen}
}
\value{
Een vector met TWN-informatie. Namen die niet in de TWN worden
gevonden krijgen \code{NA}.
}
\description{
Deze functies zoeken bij taxonnamen de bijbehorende informatie uit de
TWN-lijst.
}
\details{
Deze functies accepteren een vector met taxonnamen en retourneren de
betreffende informatie uit de TWN-lijst. Als de taxonnaam niet in de
TWN-lijst voorkomt retourneert de functie \code{NA}
\itemize{
\item \code{get_twn_voorkeur}: Geeft de voorkeurnaam van het taxon. Als het al de
voorkeurnaam is wordt de originele naam teruggegeven.
\item \code{get_twn_parent}: Geeft de parent van het taxon.
\item \code{get_twn_status}: Geeft de status van het taxon.
\item \code{get_twn_localname}: Geeft de Nederlandse naam van het taxon.
\item \code{get_twn_taxonlevel}: Geeft het taxonlevel van het taxon.
}
}
\note{
De TWN-lijst is niet altijd volledig consistent. Diverse taxa staan
meerdere malen in de lijst. In dat geval wordt het eerste resultaat
geretourneerd.

Als een taxon niet de voorkeurnaam is dan ontbreekt er vaak informatie
zoals de parent van het taxon
}
\examples{

get_twn_voorkeur(c("Bufo calamita", "Bufo"))
get_twn_parent(c("Bufo calamita", "Bufo"))
get_twn_status(c("Bufo calamita", "Bufo"))
get_twn_localname(c("Bufo calamita", "Bufo"))
get_twn_taxonlevel(c("Bufo calamita", "Bufo"))

}
\seealso{
\link[autecodata:twn_lijst]{autecodata::twn_lijst}
}
