% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_get_data.R
\name{get_ws_grens}
\alias{get_ws_grens}
\title{Waterschapsgrens}
\usage{
get_ws_grens(pad = "data")
}
\arguments{
\item{pad}{een string voor het pad op de ws_grens.gpkg op te slaan. ordt aangemaakt als het nog niet bestaat.
Default is de map \emph{data}.}
}
\value{
\code{get_ws_grens()} slaat ws_grens.gpkg op de gekozen locatie.
}
\description{
Waterschapsgrens als bestand
}
\examples{

\dontrun{
get_ws_grens("data")
}

}
