% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{gisdata_info}
\alias{gisdata_info}
\title{Informatie over Open GIS data van HHSK}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 7 columns.
}
\usage{
gisdata_info
}
\description{
Bevat een overzicht van een groot deel van de open GIS data van HHSK met de bijbehorende
URL's om de online data op te halen.
}
\keyword{datasets}
