% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{grafiek_basis}
\alias{grafiek_basis}
\title{Tijdreeksgrafiek}
\usage{
grafiek_basis(data, mp = NULL, mpomsch = NULL, parnaam = NULL,
  eenheid = NULL, plot_loess = TRUE)
}
\arguments{
\item{data}{Een dataframe met de gegevens van 1 tijdreeks, dus van 1 meetpunt en 1 parameter.
Kolommen zoals beschreven in \code{\link[=import_fys_chem]{import_fys_chem()}}.}

\item{mp}{Code van het meetpunt. Deze wordt gebruikt in de titel}

\item{mpomsch}{Omschrijving van het meetpunt. Deze wordt gebruikt in de titel}

\item{parnaam}{Naam van de parameter. Deze wordt gebruikt in de ondertitel}

\item{eenheid}{Eenheid van de parameter. Deze wordt gebruikt als titel van de Y-as}

\item{plot_loess}{Wel of niet plotten van een LOESS-curve. Default is TRUE}
}
\value{
Een ggplot grafiek. Het is mogelijk om achteraf andere ggplot objecten toe te voegen met \code{+}
}
\description{
Deze functie maakt een tijdreeksgrafiek van 1 meetpunt van 1 parameter.
}
