% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{hhskblauw}
\alias{hhskblauw}
\title{HHSK-blauw}
\format{
An object of class \code{character} of length 1.
}
\usage{
hhskblauw
}
\description{
De kleur blauw die overeenkomt met het HHSK-logo
}
\keyword{datasets}
