% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{hhskgroen}
\alias{hhskgroen}
\title{HHSK-groen}
\format{
An object of class \code{character} of length 1.
}
\usage{
hhskgroen
}
\description{
De kleur groen die overeenkomt met het HHSK-logo
}
\keyword{datasets}
