% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{hhskthema}
\alias{hhskthema}
\title{HHSK thema}
\usage{
hhskthema()
}
\value{
Een thema voor ggplot
}
\description{
Dit is een HHSK-thema voor ggplot met de kleuren van HHSK
}
\examples{
\dontrun{

my_ggplot + hhskthema()

}
}
