% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_package_description.R
\docType{package}
\name{HHSKwkl}
\alias{HHSKwkl}
\title{HHSKwkl: Een package voor waterkwaliteit}
\description{
Functies voor waterkwaliteitsanalyse. Zie verder: \url{https://redtent.github.io/HHSKwkl/}
}
