% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{import_fys_chem}
\alias{import_fys_chem}
\title{Importeren van fysisch-chemische data}
\usage{
import_fys_chem(fys_chem_csv = "data/fys_chem.csv")
}
\arguments{
\item{fys_chem_csv}{Een characterstring met het pad naar het te importeren bestand. Het bestand moet in
csv-formaat zijn met \code{;} als scheidingsteken en \code{,} als scheidingsteken. Default is \code{"data/fys_chem.csv"}.
Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.}
}
\value{
Een dataframe met fysisch-chemische meetgegevens.
}
\description{
De functie helpt bij het importeren van fysische chemische data. De functie leest de kolom \code{datum} als \code{Date}.
Verder worden rijen zonder \code{waarde} verwijderd.
}
\details{
Er is enige vrijheid t.a.v. de inhoud van het bestand. De functie verwacht ten minste een kolom
\code{datum} en \code{waarde}. Een standaard bestand heeft gewoonlijk de kolommen:
\itemize{
\item \code{mp} Code met de aanduiding van het meetpunt
\item \code{datum} Datum in het format dd-mm-yyyy hh:mm:ss
\item \code{parnr} Unieke nummer van de parameter
\item \code{par} Parametercode, voor de leesbaarheid
\item \code{eenheid} Eenheid, voor de leesbaarheid
\item \code{detectiegrens} Aanduiding als < of >. Bij geen aanduidng krijgt het veld \code{NA}
\item \code{waarde} Waarde van de meting

}
}
\examples{
\dontrun{

data <- import_fys_chem()

}
}
