% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{import_normen_rivm}
\alias{import_normen_rivm}
\title{Importeren RIVM normen}
\usage{
import_normen_rivm(normen = "data/normen.txt",
  parameterdf = import_parameters())
}
\arguments{
\item{normen}{Het normenbestand zoals dit van de RIVM website gedownload kan worden. Default is \code{"data/normen.txt"}}

\item{parameterdf}{Een parameterbestand. Zie ook \code{\link{import_parameters}}. Voor de koppeling zijn de kolommen \code{aquo_parcode},
\code{aquo_eenheid} en \code{aquo_compartiment} noodzakelijk. Default wordt een nieuw parameterbestand ingelezen met
\code{import_parameters()}}
}
\value{
Een dataframe met de kolommen \code{parnr}, \code{norm_JGM}, \code{norm_MAX}, \code{norm_P90}, \code{min_norm}
}
\description{
Het rivm stelt via de website \url{https://rvs.rivm.nl/zoeksysteem/} normen beschikbaar voor stoffen. Via deze website kunnen de
normen voor zoet oppervlaktewater worden gedownload. Deze functie leest een bestand met de normen zoals
de website deze beschikbaar stelt en koppelt deze aan Dawaco-parameternummers. De koppeling is op basis van parametercode,
eenheid en compartiment. Hoedanigheid wordt niet meegenomen. Veel normen hebben als hoedanigheid nf. Dit verschil levert echter
slechts een geringe afwijking (< 5 \%) en is voor de toetsing niet echt relevant. Alleen gewasbeschermingsmiddelen worden meegenomen.
}
\examples{
\dontrun{

normen <- import_normen_rivm()

}
}
