% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_data.R
\name{import_parameters}
\alias{import_parameters}
\title{Importeren van parameters}
\usage{
import_parameters(parameter_csv = "data/parameters.csv")
}
\arguments{
\item{parameter_csv}{Een characterstring met het pad naar het te importeren bestand. Het bestand moet in
csv-formaat zijn met \code{;} als scheidingsteken en \code{,} als scheidingsteken. Default is \code{"data/parameters.csv"}.
Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.}
}
\value{
Een dataframe met parameterinformatie

#' @details Er is enige vrijheid t.a.v. de inhoud van het bestand. Een standaard bestand heeft minimaal
de onderstaande kolommen, maar de aquo_kolommen zijn wel wenselijk.
\itemize{
\item \code{parnr} Het parameternummer
\item \code{par} Code voor de parameter, veelal aquo-conform
\item \code{parnaamlang} Uitgebreide naam van de parameter. Geschikt voor weergave
\item \code{eenheid} Eenheid van de parameter
}
}
\description{
De functie helpt bij het importeren van parameter-data en bijbehorende parameterinformatie.
}
\examples{
\dontrun{

parameters <- import_parameters()

}
}
