% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{kleuren}
\alias{kleuren}
\alias{blauw}
\alias{blauw_m}
\alias{blauw_l}
\alias{oranje}
\alias{oranje_m}
\alias{oranje_l}
\alias{grijs}
\alias{grijs_m}
\alias{grijs_l}
\title{Kleuren}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
blauw

blauw_m

blauw_l

oranje

oranje_m

oranje_l

grijs

grijs_m

grijs_l
}
\description{
Diverse kleurentinten blauw (positief), oranje (negatief), en grijs (neutraal). Toevoeging _m of _l voor
midden of lichte variant.
}
\keyword{datasets}
