% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_rapportage_utils.R
\name{link_meetgegevens}
\alias{link_meetgegevens}
\title{HTML-link naar meetpuntpagina met meetgegevens}
\usage{
link_meetgegevens(mp, link_tekst = NULL)
}
\arguments{
\item{mp}{De meetpuntcode}

\item{link_tekst}{Optionele tekst}
}
\value{
HTML-tekst met een link naar een grafiek
}
\description{
HTML-link naar meetpuntpagina met meetgegevens
}
\examples{

\dontrun{
link_meetgegevens("S_0067")
link_meetgegevens("S_0067", "eigen_tekst")

}
}
