% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{logo}
\alias{logo}
\title{HHSK logo}
\format{
An object of class \code{raster} with 184 rows and 893 columns.
}
\usage{
logo
}
\description{
Dit is het HHSK logo
}
\keyword{datasets}
