% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{maak_opzoeker}
\alias{maak_opzoeker}
\title{Maak een opzoekfunctie}
\usage{
maak_opzoeker(df, key = 1, value = 2)
}
\arguments{
\item{df}{Een dataframe met de opzoekwaarden.}

\item{key}{De sleutelkolom om een waarde op te zoeken. Default is de eerste
kolom.}

\item{value}{De kolom met waarden die opgezocht moeten worden. Default is de
tweede kolom}
}
\value{
Een functie die gebruikt kan worden voor het opzoeken van waarden. De
functie accepteert (een vector) met namen en retourneert (een vector) met
opgezochte waarden.
}
\description{
Deze functie maakt opzoekfuncties waarmee een waarde op basis van een sleutel
opgezocht kan worden.
}
\examples{

x <- USArrests
x$names <- rownames(x)

urbanpop_state <- maak_opzoeker(x, key = names, value = UrbanPop)
urbanpop_state("Texas")
urbanpop_state(c("Vermont", "Texas"))


}
