% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{maand_namen}
\alias{maand_namen}
\title{Maandnamen}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
}
\usage{
maand_namen
}
\description{
Dit is een dataframe met 3 kolommen.
\itemize{
\item \code{maand} - Integer met het maandnummer
\item \code{maand_namen} - Geordende factor met maandnamen, lowercase
\item \code{Maand_namen} - Geordende factor met maandnamen, titlecase
}
}
\keyword{datasets}
