% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_biologie_info.R
\name{macev_groep}
\alias{macev_groep}
\title{Bepaal macrofaunagroep}
\usage{
macev_groep(naam)
}
\arguments{
\item{naam}{Naam van een taxon. Als het taxon ontbreekt of niet gevonden wordt dan
is het resultaat \code{"Overig"}}
}
\value{
Een vector met de groepsnamen
}
\description{
Deze functie bepaalt de groep waartoe een macrofaunataxon behoort, De indeling is gebaseerd op
het werk van de trainee Jildau Kuiper.
}
\examples{

macev_groep(c("Faxonius", "Procambarus", "TEST", NA))

}
