% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{make_sf}
\alias{make_sf}
\title{Maak een dataframe een simple feature}
\usage{
make_sf(df, x = "x", y = "y", remove = FALSE)
}
\arguments{
\item{df}{Een dataframe}

\item{x}{Naam van de x-kolom. Default is \code{"x"}}

\item{y}{Naam van de y-kolom. Default is \code{"y"}}

\item{remove}{logical; when coords or wkt is given, remove these columns from data.frame? Default is \code{FALSE}}
}
\value{
Een simple feaure dataframe
}
\description{
Deze functie zet een dataframe om in een Simple Feature. De functie maakt gebruik van \code{\link[sf:st_as_sf]{sf::st_as_sf()}}, maar
heeft defaults die het dagelijks gebruik eenvoudiger maken. De functie gaat uit van RD-coordinaten en
een x-kolom "x" en een y-kolom "y".
}
\examples{

\dontrun{

meetpunten \%>\% make_sf()

}

}
