% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_monster_info.R
\name{mon_info}
\alias{mon_info}
\alias{mon_taxa_n}
\alias{mon_taxa_missing}
\alias{mon_taxa_n_uniek}
\alias{mon_taxa_n_dubbel}
\alias{mon_twn_niet_in_twn}
\alias{mon_twn_status_afwijkend}
\alias{mon_twn_synoniem}
\alias{mon_twn_fout}
\alias{mon_twn_aggregaat}
\alias{mon_waarde_missing}
\alias{mon_waarde_nul}
\alias{mon_waarde_som}
\alias{mon_waarde_max}
\alias{mon_level_kd_species}
\alias{mon_level_gd_species}
\title{Informatie over de taxa in het monster}
\usage{
mon_taxa_n(naam)

mon_taxa_missing(naam)

mon_taxa_n_uniek(naam)

mon_taxa_n_dubbel(naam)

mon_twn_niet_in_twn(naam)

mon_twn_status_afwijkend(naam)

mon_twn_synoniem(naam)

mon_twn_fout(naam)

mon_twn_aggregaat(naam)

mon_waarde_missing(waarde)

mon_waarde_nul(waarde)

mon_waarde_som(waarde)

mon_waarde_max(waarde)

mon_level_kd_species(naam)

mon_level_gd_species(naam)
}
\arguments{
\item{naam}{Een vector met taxonnamen.}

\item{waarde}{Een vector met aantallen (per taxon).}
}
\value{
Een enkel getal waarmee het betreffende aspect wordt samengevat.
}
\description{
Verschillende functies om de diverse aspecten van taxa en waarden in het
monster samen te vatten.
}
\details{
Iedere functie geeft een samenvatting van een aspect van de
taxonnamen van het monster.
\itemize{
\item \code{mon_taxa_missing}  - Het aantal ontbrekende taxonnamen.
\item \code{mon_taxa_n}        - Het aantal taxonnamen, mogelijk inclusief dubbelingen.
\item \code{mon_taxa_n_uniek}  - Het aantal unieke taxonnamen.
\item \code{mon_taxa_n_dubbel} - Het aantal taxa dat 2 keer of vaker voorkomt.
\item \code{mon_twn_niet_in_twn}      - Aantal taxa dat niet in de TWN-lijst voorkomt.
\item \code{mon_twn_status_afwijkend} - Aantal taxa dat niet status 10 (voorkeursnaam) heeft.
\item \code{mon_twn_synoniem}         - Aantal taxa dat een synoniem is (status 20).
\item \code{mon_twn_aggregaat}        - Aantal taxa dat een aggregaat van verschillende soorten is (status 80).
\item \code{mon_twn_fout}             - Aantal taxa dat een foutcode heeft in de TWN-lijst (status 30, 91, en 92)
\item \code{mon_waarde_missing} - Aantal ontbrekende waarden.
\item \code{mon_waarde_nul}     - Aantal waarden die (kleiner dan) 0 zijn.
\item \code{mon_waarde_tot}     - Som van alle waarden in het monster.
\item \code{mon_waarde_max}     - Hoogst voorkomende waarde in het monster.
\item \code{mon_level_kd_species} - Aantal taxa op een lager niveau dan \emph{Species}
\item \code{mon_level_gd_species} - Aantal taxa op een hoger niveau dan \emph{Species}
}
}
\note{
Als taxa meerdere keren in het monster voorkomen dan worden deze in de \code{mon_twn_*}-functies
ook meerdere malen geteld.
}
\examples{

mon_taxa_missing(c("Bufo calamita", "Bufo", NA, "Bufo"))
mon_taxa_n(c("Bufo calamita", "Bufo", NA, "Bufo"))
mon_taxa_n_uniek(c("Bufo calamita", "Bufo", NA, "Bufo"))
mon_taxa_n_dubbel(c("Bufo calamita", "Bufo", NA, "Bufo"))

mon_twn_niet_in_twn(c("Bufo calamita", "Bufo", NA, "Bufo", "Denkbeeldig dier"))
mon_twn_status_afwijkend(c("Bufo calamita", "Bufo", NA, "Bufo", "Denkbeeldig dier"))
mon_twn_synoniem(c("Bufo calamita", "Bufo", NA, "Bufo", "Denkbeeldig dier"))
mon_twn_aggregaat(c("Bufo calamita", "Bufo", NA, "Bufo", "Denkbeeldig dier"))
mon_twn_fout(c("Bufo calamita", "Bufo", NA, "Bufo", "Denkbeeldig dier"))

mon_waarde_missing(c(0, 1, 5, 20, 50, NA, NA))
mon_waarde_nul(c(0, 1, 5, 20, 50, NA, NA))
mon_waarde_som(c(0, 1, 5, 20, 50, NA, NA))
mon_waarde_max(c(0, 1, 5, 20, 50, NA, NA))
}
