% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_monster_info.R
\name{mon_summary}
\alias{mon_summary}
\title{Samenvatting monster}
\usage{
mon_summary(df, naam = naam, waarde = waarde)
}
\arguments{
\item{df}{Een (grouped) dataframe met een of meerdere monsters}

\item{naam}{De kolomnaam met de taxonnamen. Dit argument wordt gequote.}

\item{waarde}{De kolomnaam met de waarden. Dit argument wordt gequote.}
}
\value{
DUMMY
}
\description{
Deze functie maakt een samenvatting van een monster. Hiermee is
het mogelijk om snel een overzicht van een monster te krijgen en eventuele
problemen te constateren.

Het is mogelijk om meerdere monsters tegelijk samen te vatten door gebruik
van de functie \link[dplyr:group_by]{dplyr::group_by}
}
\examples{

#DUMMY
}
