% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_verwerking_data.R
\name{mspaf}
\alias{mspaf}
\title{Bereken de msPAF}
\usage{
mspaf(paf_vector)
}
\arguments{
\item{paf_vector}{Een vector met PAF-waarden (tussen de 0 en 1).}
}
\value{
De msPAF
}
\description{
Berekent de msPAF door het effect van afzonderlijke PAF-waarden samen te nemen.
De msPAF berekening werkt volgens het principe van kansrekening.
}
\examples{

# dobbelsteen voorbeeld
mspaf(c(1/6, 1/6))

}
