% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{my_c}
\alias{my_c}
\title{Create a single clean vector}
\usage{
my_c(vector)
}
\arguments{
\item{vector}{A vector that needs 'cleaning'}
}
\value{
Returns a single clean vector without names
}
\description{
A helper function for the basic \code{c()} function to remove names and create a singular vector.
This can be useful if you select a single column from a tibble.
}
