% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{norm_lijnen}
\alias{norm_lijnen}
\alias{add_norm_lijnen}
\title{Maak en voeg RIVM-normlijnen toe aan een plot}
\usage{
norm_lijnen(parnr, normen)

add_norm_lijnen(plot, parnr, normen)
}
\arguments{
\item{parnr}{Het nummer van de parameter waar de normen van toegevoegd moeten worden}

\item{normen}{Een dataframe met de normen zoals geimporteerd kan worden met de functie \code{\link{import_normen_rivm}}}

\item{plot}{Het plot waar de normlijnen aan toegevoegd moeten worden}
}
\value{
een geom-object met normen dat toegevoegd kan worden aan een plot, of een plot waar deze aan toegevoegd zijn.
}
\description{
De functie \code{norm_lijnen} vertaalt de normen behorend bij een bepaald parameternummer in normlijnen die toegevoegd kunnen worden aan
een plot. De functie \code{add_norm_lijnen} voegt ze toe aan het plot.
}
\section{Functions}{
\itemize{
\item \code{norm_lijnen()}: Creeer normlijnen die aan een plot toegevoegd kunnen worden

\item \code{add_norm_lijnen()}: Voeg normlijnen toe aan een bestaand plot

}}
\examples{

\dontrun{

norm_lijnen_object <- norm_lijnen(parnr = 1000, normen = normen)
plot_met_normen <- plot + norm_lijnen_object

plot_met_normen <- add_norm_lijnen(plot, parnr = 1000, normen = normen )

plot \%>\% add_norm_lijnen(parnr = 1000, normen = normen)

}

}
