% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{opzoeken_waarde}
\alias{opzoeken_waarde}
\title{Zoek een waarde op in een opzoektabel}
\usage{
opzoeken_waarde(sleutel, df, attribuut, sleutelkolom = 1)
}
\arguments{
\item{sleutel}{De sleutelwaarde, selecteert de rij}

\item{df}{Een dataframe die functioneert als opzoektabel}

\item{attribuut}{De kolom(index) waar de op te halen waarde staat}

\item{sleutelkolom}{Optioneel - de kolom(index) als character of integer.
Standaard staat de eerste kolom als sleutelkolom ingesteld}
}
\value{
De waarde die te vinden is op de betreffende rij of kolom.
Bij dubbelingen in de sleutel, wordt de eerste waarde geretourneerd.
}
\description{
Deze functie kan waarden opzoeken in een tabel. Het is het alternatief voor een hash table, die R helaas niet heeft.
}
\examples{
\dontrun{

meetpuntomschrijving <- ophalen_waarde(sleutel = "00016", df = meetpuntendf, attribuut = "mpomsch")

}
}
