% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_verwerking_data.R
\name{paf_gbm}
\alias{paf_gbm}
\title{Bereken PAF voor bestrijdingsmiddelen}
\source{
De toxiciteitsdata komen uit de \href{https://rivm.shinyapps.io/chemie-tool-in-shiny/}{toxiciteitstool}.
Documentatie is te vinden op de
\href{https://kiwk-tox.netlify.app/nl/chemie-tool-rekentool/}{website van ESF 8 Toxiciteit}
}
\usage{
paf_gbm(
  aquo_par,
  concentratie,
  detectiegrens,
  ssd_data,
  type_paf = c("acuut", "chronisch"),
  digits = NULL
)
}
\arguments{
\item{aquo_par}{Parameternaam volgens de aquonaamgeving}

\item{concentratie}{Concentratie in ug/l}

\item{detectiegrens}{Detectiegrens. NA als geen detectiegrens (NB TRUE FALSE zou logischer zijn)}

\item{ssd_data}{SSD data volgens het format van de toxiciteitstool}

\item{type_paf}{\code{"acuut"} of \code{"chronisch"}}

\item{digits}{\code{NULL} of het aantal decimalen om de paf af te ronden}
}
\value{
Een vector met PAF waarden
}
\description{
Functie voor het berekenen van potentieel aangetaste fractie obv de data van de
toxiciteitstool van ESF 8 Toxiciteit versie 2.
}
\examples{

\dontrun{
testdata \%>\% 
mutate(paf_acuut = paf_gbm(aquo_par = f_aquopar(parnr), 
                           concentratie = waarde, 
                           detectiegrens = detectiegrens,
                           ssd_data = tox, type_paf = "acuut"),
       paf_chronisch = paf_gbm(aquo_par = f_aquopar(parnr), 
                               concentratie = waarde, 
                               detectiegrens = detectiegrens,
                               ssd_data = tox, type_paf = "chronisch", digits = 5))

}
}
