% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_dataframe_utils.R
\name{randomize}
\alias{randomize}
\title{Randomize}
\usage{
randomize(data)
}
\arguments{
\item{data}{Een dataframe}
}
\value{
Hetzelfde dataframe als de input in willekeurige volgorde
}
\description{
Verandert de rijen in willekeurige volgorde
}
\examples{
\dontrun{

randomize(mtcars)
}
}
