% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_render_server.R
\name{render_server}
\alias{render_server}
\title{Render R-markdown documenten op de server}
\usage{
render_server(rmd_bestand, rmd_map, output_bestand, output_map = rmd_map,
  temp_dir = "C:/JT", ...)
}
\arguments{
\item{rmd_bestand}{Naam van het R-markdownbestand}

\item{rmd_map}{Pad van het R-markdownbestand}

\item{output_bestand}{Naam van het outputbestand}

\item{output_map}{Pad naar de outputlocatie}

\item{temp_dir}{Naam van de tijdelijke map. Default is \code{"C:/JT"}}

\item{...}{Optionele argumenten naar \code{rmarkdown::render}}
}
\value{
Een gerenderd R-markdownbestand
}
\description{
De R-server heeft problemen om op het netwerk te renderen. Deze functie omzeilt het
probleem door R-markdown documenten eerst lokaal te renderen en daarna to kopieren
naar de gewenste locatie.
}
\note{
Deze functie is een "hack" om het renderen op de server zonder problemen te
laten verlopen. Het zou wenselijker zijn als direct de functie \code{rmarkdown::render()}
kan worden aangeroepen.
}
