% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{schonevoeten}
\alias{schonevoeten}
\title{Droge voeten en schoon water}
\format{An object of class \code{raster} with 148 rows and 627 columns.}
\usage{
schonevoeten
}
\description{
Beeld Droge voeten en schoon water
}
\keyword{datasets}
