% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{titelpagina_internet}
\alias{titelpagina_internet}
\title{Titelpagina voor de grafieken op internet}
\usage{
titelpagina_internet(inclusief_normen = TRUE)
}
\arguments{
\item{inclusief_normen}{logical Bepaald of de lijnen met normen op de titelpagina worden weergegeven. Default is \code{TRUE}}
}
\value{
Een titelpagina voor de grafieken
}
\description{
Creeert een titelpagina voor grafieken zoals deze worden geproduceerd met \link{grafiek_basis}. Dient gebruikt te worden binnen een graphic device (bijvoorbeeld \code{pdf()}).
}
\examples{
\dontrun{

pdf("mijntest.pdf", width = 16, height = 8)
titelpagina_internet()
print(basisgrafiek)
dev.off()
}
}
