% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_verwerking_data.R
\name{toetsing_gbm}
\alias{toetsing_gbm}
\title{Toetsing gewasbeschermingsmiddelen}
\usage{
toetsing_gbm(data, normen, factor_detectiegrens = 0.5)
}
\arguments{
\item{data}{Een dataframe met data zoals wordt gemaakt met \code{\link{import_fys_chem}}}

\item{normen}{Een dataframe met normen zoals wordt gemaakt met \code{\link{import_normen_rivm}}}

\item{factor_detectiegrens}{Te gebruiken factor voor detectiegrenswaarden. Default is 0.5.}
}
\value{
Een dataframe met per meetpunt, parameter en jaar een toetsing. De toetsing bestaat uit:
\itemize{
\item \code{aangetroffen} - Logical. Of het gewasbeschermingsmiddel is aangetroffen of niet
\item \code{overschrijding} - Logical. Of er een normoverschrijding geconstateerd is of niet
\item \code{hoogste_overschrijding} - De maximale (overschrijdings)factor
\item \code{factor_MAX} - De (overschrijdings)factor voor de Maximaal Aanvaarbare Concentratie - MAC
\item \code{facor_JGM} - De (overschrijdings)factor voor de jaargemiddelde concentratie - JGM
\item \code{factor_P90} - De (overschrijdings)factor voor de 90-percentielwaarde - P90
}
}
\description{
Deze functie toetst per jaar per gewasbeschermingsmiddel of er een overschrijding is en hoe groot deze is.
}
\examples{
\dontrun{

toetsing <- toetsing_gbm(mijn_data, normen)
}

}
