% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_get_twn_info.R
\name{twn_info}
\alias{twn_info}
\alias{fun_twn_voorkeur}
\alias{fun_twn_parent}
\alias{fun_twn_status}
\alias{fun_twn_localname}
\alias{fun_twn_taxonlevel}
\alias{get_twn_voorkeur}
\alias{get_twn_parent}
\alias{get_twn_status}
\alias{get_twn_localname}
\alias{get_twn_taxonlevel}
\title{Opzoeken van de TWN informatie}
\usage{
fun_twn_voorkeur(namen)

fun_twn_parent(namen)

fun_twn_status(namen)

fun_twn_localname(namen)

fun_twn_taxonlevel(namen)

get_twn_voorkeur(namen)

get_twn_parent(namen)

get_twn_status(namen)

get_twn_localname(namen)

get_twn_taxonlevel(namen)
}
\arguments{
\item{namen}{Een vector met taxonnamen.}
}
\value{
Een vector met TWN-informatie. Namen die niet in de TWN worden
gevonden krijgen \code{NA}.
}
\description{
De \code{fun_twn_*}-functies maken een opzoekfunctie om bij taxonnamen de bijbehorende informatie uit de
TWN-lijst op te zoeken. De \code{get_twn_*}-functies gebruiken deze opzoekfuncties om direct deze informatie
uit de TWN-lijst op te halen. Voor herhaaldelijk gebruik zijn de \code{get_twn_*}-functies echter minder efficient.
}
\details{
Deze functies accepteren een vector met taxonnamen en retourneren de
betreffende informatie uit de TWN-lijst. Als de taxonnaam niet in de
TWN-lijst voorkomt retourneert de functie \code{NA}
\itemize{
\item \code{get_twn_voorkeur}: Geeft de voorkeurnaam van het taxon. Als het al de
voorkeurnaam is wordt de originele naam teruggegeven.
\item \code{get_twn_parent}: Geeft de parent van het taxon.
\item \code{get_twn_status}: Geeft de status van het taxon.
\item \code{get_twn_localname}: Geeft de Nederlandse naam van het taxon.
\item \code{get_twn_taxonlevel}: Geeft het taxonlevel van het taxon.
}
}
\note{
De TWN-lijst is niet altijd volledig consistent. Diverse taxa staan
meerdere malen in de lijst. In dat geval wordt het eerste resultaat
geretourneerd.

Als een taxon niet de voorkeurnaam is dan ontbreekt er vaak informatie
zoals de parent van het taxon.
}
\examples{
get_twn_voorkeur2   <- fun_twn_voorkeur()
get_twn_parent2     <- fun_twn_parent()
get_twn_status2     <- fun_twn_status()
get_twn_localname2  <- fun_twn_localname()
get_twn_taxonlevel2 <- fun_twn_taxonlevel()


get_twn_voorkeur(c("Bufo calamita", "Bufo"))
get_twn_parent(c("Bufo calamita", "Bufo"))
get_twn_status(c("Bufo calamita", "Bufo"))
get_twn_localname(c("Bufo calamita", "Bufo"))
get_twn_taxonlevel(c("Bufo calamita", "Bufo"))

}
\seealso{
Deze functies werken op basis van de tabel \link[autecodata:twn_lijst]{autecodata::twn_lijst}
}
