% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{waterschapsgrens}
\alias{waterschapsgrens}
\alias{ws_grens_rd}
\alias{ws_grens_wgs}
\title{Waterschapsgrens}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
ws_grens_rd

ws_grens_wgs
}
\description{
Waterschapsgrens als sf-object in RD-stelstel (CRS 28992) en WGS84 (CRS 4326)
}
\keyword{datasets}
