% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_get_data.R
\name{ws_grens}
\alias{ws_grens}
\alias{get_ws_grens}
\title{Waterschapsgrens}
\usage{
ws_grens(crs = NULL)

get_ws_grens(pad = "data")
}
\arguments{
\item{crs}{optionele CRS-code als getal voor conversie van de CRS}

\item{pad}{een string voor het pad op de ws_grens.gpkg op te slaan. Default is de map \emph{data}.}
}
\value{
\code{ws_grens()} geeft een sf dataframe met de waterschapsgrens.

\code{get_ws_grens()} slaat ws_grens.gpkg op de gekozen locatie.
}
\description{
Waterschapsgrens als dataframe of als bestand
}
\examples{

\dontrun{
ws_grens(crs = 4326)

get_ws_grens("data")}

}
