% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_kaarten.R
\name{basiskaart}
\alias{basiskaart}
\title{Leaflet basiskaart met topo en luchtfoto}
\usage{
basiskaart(data = NULL)
}
\arguments{
\item{data}{Een optioneel dataframe. Zie ook het data-argument in de functie \code{\link[leaflet:leaflet]{leaflet::leaflet()}}}
}
\value{
Een leaflet kaart
}
\description{
Deze functie genereert een basiskaart in leaflet. Aan deze kaart kunnen vervolgens vrij elementen worden toegevoegd
}
\examples{

basiskaart()

}
