% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_kaarten.R
\name{basiskaart}
\alias{basiskaart}
\title{Leaflet basiskaart met topo en luchtfoto}
\usage{
basiskaart(data = NULL)
}
\arguments{
\item{data}{Een optioneel dataframe. Dataframe wordt doorgeleid naar de fucntie \code{\link[leaflet:leaflet]{leaflet::leaflet()}}}
}
\description{
Leaflet basiskaart met topo en luchtfoto
}
