% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_grafieken.R
\name{boxplot_basis}
\alias{boxplot_basis}
\title{Boxplots per jaar}
\usage{
boxplot_basis(data, meetpuntendf = import_meetpunten(),
  parameterdf = import_parameters(), mp = NULL, parnr = NULL, ...)
}
\arguments{
\item{data}{Een dataframe met de gegevens van 1 tijdreeks, dus van 1 meetpunt en 1 parameter.
Kolommen zoals beschreven in \code{\link{import_fys_chem}}.}

\item{meetpuntendf}{dataframe. Een opzoektabel voor de locatie-omschrijving.
Kolommen zoals beschreven in \code{\link{import_meetpunten}}.
Probeert default ook met deze functie een meetpuntendf te maken.}

\item{parameterdf}{dataframe. Een opzoektabel voor de uitgebreide parameternaam en eenheid.
Kolommen zoals beschreven in \code{\link{import_parameters}}.
Probeert default ook met deze functie een parameterdf te maken.}

\item{mp}{character. Optioneel, Meetpuntcode van het betreffende meetpunt. Neemt anders het eerste meetpunt uit \code{data}}

\item{parnr}{character. Optioneel,Parameternummer van het betrffende meetpunt. Neemt anders het eerste parameternummer uit \code{data}}

\item{...}{Heeft geen functie, maar kan gebruikt worden om overbodige parameters te negeren}
}
\value{
Een ggplot boxplot grafiek. Het is mogelijk om achteraf andere ggplot objecten toe te voegen met \code{+}
}
\description{
Deze functie plot een boxplot per jaar van 1 meetpunt van 1 parameter. De functie dient eigenlijk alleen met complete meetjaren te worden gebruikt.
}
\examples{
\dontrun{

basis_boxplot_grafiek <- boxplot_basis(data = chloride_myplace, 
                    parameterdf, meetpuntendf) 
}
}
