% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_import_utils.R
\name{copy_data}
\alias{copy_data}
\title{Synchroniseer bestanden met een bronlocatie}
\usage{
copy_data(filenames = NULL, origin, destination = "data")
}
\arguments{
\item{filenames}{Vector of lijst met bestanden. Als deze leeg blijft worden
alle bestanden uit \code{origin} gekopieerd.}

\item{origin}{Bronlocatie.}

\item{destination}{Doellocatie. Default is \code{"data"}}
}
\value{
Synchronisatie van bestanden van de bronlocatie met de doellocatie.
}
\description{
Deze functie is bedoeld om data van een centrale locatie te synchroniseren met een projectmap.
}
\examples{

\dontrun{

copy_data()
copy_data("fys_chem.csv")
copy_data("fys_chem.csv", origin = "P:/Dawaco/FME", destination = "data")
}



}
