% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_ggplot_themas.R
\name{hhskthema}
\alias{hhskthema}
\alias{hhskthema_kaart}
\title{HHSK thema}
\usage{
hhskthema()

hhskthema_kaart()
}
\value{
Een thema voor ggplot
}
\description{
Dit is een HHSK-thema voor ggplot
}
\examples{
\dontrun{

my_ggplot + hhskthema()

}

}
