% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_ggplot_themas.R
\docType{data}
\name{hhskthema}
\alias{hhskthema}
\alias{hhskthema_kaart}
\alias{thema_hor_bar}
\alias{thema_vert_bar}
\alias{thema_line_facet}
\alias{thema_test}
\title{HHSK thema}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.

An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.

An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
hhskthema()

hhskthema_kaart()

thema_hor_bar

thema_vert_bar

thema_line_facet

thema_test()
}
\value{
Een thema voor ggplot
}
\description{
Dit is een HHSK-thema voor ggplot
}
\section{Fonts}{
Voor het correct weergeven van fonts in Rstudio of documenten moeten extra stappen
worden genomen. Zie \url{https://systemfonts.r-lib.org/articles/systemfonts.html} voor een
toelichting.
}

\examples{
\dontrun{

my_ggplot + hhskthema()

}

}
\keyword{datasets}
