% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_ggplot_themas.R
\docType{data}
\name{hhskthema}
\alias{hhskthema}
\alias{hhskthema_kaart}
\alias{thema_hor_bar}
\alias{thema_vert_bar}
\alias{thema_line_facet}
\title{HHSK thema}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.

An object of class \code{theme} (inherits from \code{gg}) of length 93.

An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
hhskthema()

hhskthema_kaart()

thema_hor_bar

thema_vert_bar

thema_line_facet
}
\value{
Een thema voor ggplot
}
\description{
Dit is een HHSK-thema voor ggplot
}
\examples{
\dontrun{

my_ggplot + hhskthema()

}

}
\keyword{datasets}
