% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_ggplot_themas.R
\name{hhskthema}
\alias{hhskthema}
\alias{hhskthema_oud}
\alias{hhskthema_kaart_oud}
\title{HHSK thema}
\usage{
hhskthema_oud()

hhskthema_kaart_oud()
}
\value{
Een thema voor ggplot
}
\description{
Dit is een HHSK-thema voor ggplot
}
\section{Fonts}{
Voor het correct weergeven van fonts in Rstudio of documenten moeten extra stappen
worden genomen. Zie \url{https://systemfonts.r-lib.org/articles/systemfonts.html} voor een
toelichting.
}

\examples{
\dontrun{

my_ggplot + hhskthema()

}

}
