% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_package_description.R
\docType{package}
\name{HHSKwkl-package}
\alias{HHSKwkl}
\alias{HHSKwkl-package}
\title{HHSKwkl: Een package voor waterkwaliteit}
\description{
Functies voor waterkwaliteitsanalyse. Zie verder: \url{https://redtent.github.io/HHSKwkl/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RedTent/HHSKwkl}
  \item Report bugs at \url{https://github.com/RedTent/HHSKwkl/issues}
}

}
\author{
\strong{Maintainer}: Johan van Tent \email{j.van.tent@hhsk.nl}

}
