% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{kleuren}
\alias{kleuren}
\alias{kleuren_hhsk}
\alias{blauw_d}
\alias{blauw}
\alias{blauw_m}
\alias{blauw_l}
\alias{oranje}
\alias{oranje_m}
\alias{oranje_l}
\alias{grijs_d}
\alias{grijs}
\alias{grijs_m}
\alias{grijs_l}
\alias{groen_d}
\alias{groen}
\alias{groen_m}
\alias{groen_l}
\alias{wit}
\alias{tekstgrijs}
\alias{blauw_m_orig}
\alias{blauw_l_orig}
\alias{groen_m_orig}
\alias{groen_l_orig}
\alias{oranje_orig}
\alias{oranje_m_orig}
\title{Kleuren HHSK}
\format{
character

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
kleuren_hhsk

blauw_d

blauw

blauw_m

blauw_l

oranje

oranje_m

oranje_l

grijs_d

grijs

grijs_m

grijs_l

groen_d

groen

groen_m

groen_l

wit

tekstgrijs

blauw_m_orig

blauw_l_orig

groen_m_orig

groen_l_orig

oranje_orig

oranje_m_orig
}
\description{
De kleuren van de huisstijl van HHSK. Gebruik de toevoeging _d, _m of _l voor donkere, midden of lichte variant.
De originele lichtere kleurvarianten zijn niet altijd geschikt voor visualisaties.
Deze zijn opgenomen met de toevoeging _orig.
}
\keyword{datasets}
