% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_description.R
\docType{data}
\name{kleuren}
\alias{kleuren}
\alias{kleuren_hhsk}
\alias{blauw_d}
\alias{blauw}
\alias{blauw_m}
\alias{blauw_l}
\alias{oranje}
\alias{oranje_m}
\alias{grijs_d}
\alias{grijs}
\alias{grijs_m}
\alias{grijs_l}
\alias{groen_d}
\alias{groen}
\alias{groen_m}
\alias{groen_l}
\alias{wit}
\alias{tekstgrijs}
\title{Kleuren HHSK}
\format{
character

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
kleuren_hhsk

blauw_d

blauw

blauw_m

blauw_l

oranje

oranje_m

grijs_d

grijs

grijs_m

grijs_l

groen_d

groen

groen_m

groen_l

wit

tekstgrijs
}
\description{
De kleuren van de huisstijl van HHSK. Gebruik de toevoeging _d, _m of _l voor donkere, midden of lichte variant.
}
\keyword{datasets}
